+++ server/src/immich/main.ts
@@ -4,11 +4,11 @@ import { json } from 'body-parser';
 import cookieParser from 'cookie-parser';
 import { existsSync } from 'node:fs';
 import sirv from 'sirv';
+import { excludePaths } from 'src/config';
 import { WEB_ROOT, envName, isDev, serverVersion } from 'src/domain/domain.constant';
 import { AppModule } from 'src/immich/app.module';
 import { AppService } from 'src/immich/app.service';
 import { useSwagger } from 'src/immich/app.utils';
-import { excludePaths } from 'src/infra/infra.config';
 import { otelSDK } from 'src/infra/instrumentation';
 import { ImmichLogger } from 'src/infra/logger';
 import { WebSocketAdapter } from 'src/infra/websocket.adapter';
