+++ server/src/infra/instrumentation.ts
@@ -13,9 +13,9 @@ import { snakeCase, startCase } from 'lodash';
 import { OpenTelemetryModuleOptions } from 'nestjs-otel/lib/interfaces';
 import { copyMetadataFromFunctionToFunction } from 'nestjs-otel/lib/opentelemetry.utils';
 import { performance } from 'node:perf_hooks';
+import { excludePaths } from 'src/config';
 import { DecorateAll } from 'src/decorators';
 import { serverVersion } from 'src/domain/domain.constant';
-import { excludePaths } from 'src/infra/infra.config';
 
 let metricsEnabled = process.env.IMMICH_METRICS === 'true';
 const hostMetrics =
